+++ table/tables/bounded_tables.go
@@ -161,7 +161,7 @@ func (t *BoundedTable) RecordKey(h int64, col *table.Column) kv.Key {
 	if col != nil {
 		colID = col.ID
 	}
-	return encodeRecordKey(t.recordPrefix, h, colID)
+	return tablecodec.EncodeRecordKey(t.recordPrefix, h, colID)
 }
 
 // FirstKey implements table.Table FirstKey interface.
